refactor(hooks): split useQueries.ts into focused modules + centralized logger
- usePublicQueries.ts  — public hooks (menu, events, news, products, tracks, leaderboard)
- useAuthenticatedQueries.ts — JWT-protected hooks (profile, orders, cart, gamipress)
- useMutations.ts      — all write operations (cart, profile, newsletter, subscription, tracking)
- useQueries.ts        — barrel re-export for zero breaking changes in 15 existing consumers
- src/lib/logger.ts    — structured logger replacing 7 ad-hoc console.error/info calls;
                         dev = human-readable, prod = JSON ready for any log drain
